Add state management and comments in controllers
Added comments and state management code (ViewData, ViewBag, TempData) in the `StudentController` class. Updated the `Index` action method of `StudentController` to store the title "Student List" using `ViewBag.Title` and `ViewData["Title"]`. Enhanced the `Register` action method of `StudentController` to store a value in `TempData["test"]` for the next request and included a commented-out redirect to the `Teacher` controller's `Index` action. Modified the `Index` action method of `TeacherController` to check and retrieve the value of `TempData["test"]`. Updated the `Index.cshtml` view to display the `ViewBag.Title` and `ViewData["Title"]` values.